Add multi-letter information to a motif.

If the original sequences are available for a particular motif, then they can be used to generate higher-order PPM matrices.

add_multifreq(motif, sequences, add.k = 2:3, RC = FALSE,
  threshold = 0.01, threshold.type = "logodds", motifs.perseq = 1)

Arguments

motif
See convert_motifs() for acceptable formats. If the motif is not a universalmotif motif, then it will be converted.

sequences
Biostrings::XStringSet The alphabet must match that of the motif. If these sequences are all the same length as the motif, then they are all used to generate the multi-freq matrices. Otherwise scan_sequences() is first run to find the right sequence.

add.k
numeric(1) The k-let lengths to add.

RC
logical(1) Check the reverse complement of a DNA sequence. See scan_sequences().

threshold
numeric(1) Between 0 and 1. See scan_sequences().

threshold.type
character(1) One of c('logodds', 'pvalue'). See scan_sequences().

motifs.perseq
numeric(1) If scan_sequences() is run, then this indicates how many hits from each sequence is to be used.

Value

A universalmotif object with filled multifreq slot.

Details

At each position in the motif, then the probability of each k-let covering from the initial position to ncol - 1 is calculated. Only positions within the motif are considered; this means that the final k-let probability matrix will have ncol - 1 fewer columns. Calculating k-let probabilities for the missing columns would be trivial however, as you would only need the background frequencies. Since these would not be useful for scan_sequences() though, they are not calculated.

Currently add_multifreq() does not try to stay faithful to the default motif matrix when generating multifreq matrices. This means that if the sequences used for training are completely different from the actual motif, the multifreq matrices will be as well. However this is only really a problem if you supply add_multifreq() with a set of sequences of the same length as the motif; in this case add_multifreq() is forced to create the multifreq matrices from these sequences. Otherwise add_multifreq() will scan the input sequences for the motif and use the best matches to construct the multifreq matrices.

This 'multifreq' representation is only really useful within the universalmotif enrivonment. Despite this, if you wish it can be preserved in text using write_motifs().

Note: the number of rows for each k-let matrix is n^k, with n being the number of letters in the alphabet being used. This means that the size of the k-let matrix can become quite large as k increases. For example, if one were to wish to represent a DNA motif of length 10 as a 10-let, this would require a matrix with 1,048,576 rows (though at this point if what you want is to search for exact sequence matches, the motif format itself is not very useful).

See also

Examples

sequences <- create_sequences(seqlen = 10)
motif <- create_motif()
motif.trained <- add_multifreq(motif, sequences, add.k = 2:4)
## peek at the 2-let matrix:
motif.trained["multifreq"]$`2`

#>       1    2    3    4    5    6    7    8    9
#> AA 0.08 0.04 0.03 0.07 0.04 0.03 0.06 0.07 0.08
#> AC 0.09 0.10 0.05 0.07 0.10 0.05 0.08 0.09 0.08
#> AG 0.07 0.08 0.02 0.06 0.10 0.06 0.04 0.01 0.05
#> AT 0.07 0.06 0.04 0.05 0.09 0.07 0.10 0.09 0.07
#> CA 0.09 0.01 0.11 0.12 0.05 0.12 0.04 0.10 0.03
#> CC 0.05 0.08 0.06 0.04 0.07 0.03 0.05 0.05 0.06
#> CG 0.02 0.05 0.05 0.09 0.06 0.07 0.06 0.05 0.09
#> CT 0.03 0.09 0.13 0.03 0.03 0.07 0.06 0.07 0.08
#> GA 0.04 0.02 0.04 0.06 0.09 0.08 0.08 0.05 0.06
#> GC 0.03 0.08 0.08 0.02 0.08 0.07 0.07 0.09 0.03
#> GG 0.06 0.02 0.04 0.07 0.07 0.07 0.05 0.03 0.01
#> GT 0.14 0.08 0.07 0.03 0.07 0.04 0.06 0.05 0.04
#> TA 0.07 0.07 0.07 0.08 0.03 0.05 0.08 0.06 0.10
#> TC 0.06 0.09 0.09 0.08 0.04 0.06 0.07 0.03 0.09
#> TG 0.05 0.08 0.07 0.09 0.03 0.06 0.07 0.05 0.06
#> TT 0.05 0.05 0.05 0.04 0.05 0.07 0.03 0.11 0.07