Run MEME from within R.

De novo motif discovery via MEME. For a detailed description of the command, see http://meme-suite.org/doc/meme.html. For a brief description of the command parameters, run run_meme(). Parameters in run_meme() which are directly taken from the MEME program are tagged with [MEME].

run_meme(target.sequences, output = NULL, overwrite.dir = FALSE,
  control.sequences = NULL, weights = NULL, text = FALSE,
  brief = 1000, objfun = "classic", test = NULL, use_llr = FALSE,
  shuf = 2, hsfrac = NULL, cefrac = NULL, searchsize = NULL,
  norand = FALSE, csites = 1000, seed = 0, alph = NULL,
  revcomp = FALSE, pal = FALSE, mod = "zoops", nmotifs = 3,
  evt = NULL, nsites = NULL, minsites = NULL, maxsites = NULL,
  wnsites = 0.8, w = NULL, minw = 8, maxw = 50, allw = NULL,
  nomatrim = FALSE, wg = 11, ws = 1, noendgaps = FALSE,
  bfile = NULL, markov_order = 0, psp = NULL, maxiter = 50,
  distance = 0.001, prior = NULL, b = NULL, plib = NULL,
  spfuzz = NULL, spmap = NULL, cons = NULL, p = NULL,
  maxsize = NULL, maxtime = NULL, wd = getwd(),
  logfile = paste0(wd, "/memerun.log"), readsites = TRUE,
  echo = FALSE, verbose = 1, timeout = Inf,
  bin = getOption("meme.bin"))

Arguments

target.sequences
XStringSet List of sequences to get motifs from.

output
character(1) Name of the output folder. If NULL, MEME output will be deleted.

overwrite.dir
logical(1) If output is set but already exists, allow overwritting.

control.sequences
XStringSet List of negative sequences. Only used if objfun = c("de", "se").

weights
numeric Vector of numbers between 0 and 1, representing sequence weights.

text
logical(1) [MEME]

brief
numeric(1) [MEME]

objfun
character(1) [MEME]

test
character(1) [MEME]

use_llr
logical(1) [MEME]

shuf
numeric(1) [MEME]

hsfrac
numeric(1) [MEME]

cefrac
numeric(1) [MEME]

searchsize
numeric(1) [MEME]

norand
logical(1) [MEME]

csites
numeric(1) [MEME]

seed
numeric(1) [MEME]

alph
character(1) [MEME]

revcomp
logical(1) [MEME]

pal
logical(1) [MEME]

mod
character(1) [MEME]

nmotifs
numeric(1) [MEME]

evt
numeric(1) [MEME]

nsites
numeric(1) [MEME]

minsites
numeric(1) [MEME]

maxsites
numeric(1) [MEME]

wnsites
numeric(1) [MEME]

w
numeric(1) [MEME]

minw
numeric(1) [MEME]

maxw
numeric(1) [MEME]

allw
numeric(1) [MEME]

nomatrim
logical(1) [MEME]

wg
numeric(1) [MEME]

ws
numeric(1) [MEME]

noendgaps
logical(1) [MEME]

bfile
character(1) [MEME]

markov_order
numeric(1) [MEME]

psp
character(1) [MEME]

maxiter
numeric(1) [MEME]

distance
numeric(1) [MEME]

prior
character(1) [MEME]

b
numeric(1) [MEME]

plib
character(1) [MEME]

spfuzz
numeric(1) [MEME]

spmap
character(1) [MEME]

cons
character(1) [MEME]

p
numeric(1) [MEME]

maxsize
numeric(1) [MEME]

maxtime
numeric(1) [MEME]

wd
character(1) Working directory to run MEME in.

logfile
character(1) File to dump MEME stderr. If NULL, no logs will be saved.

readsites
logical(1) Read sites from MEME output (from read_meme()).

echo
logical(1) Dump MEME output to console.

verbose
numeric(1) Set verbose = 0 to quiet run_meme().

timeout
numeric(1) Stop MEME program past timeout (seconds). See processx::run().

bin
character(1) Location of MEME binary. Alternatively, set this via options(meme.bin = '/path/to/meme/bin').

Value

list The output file is read with read_meme().

References

Bailey T, Elkan C (1994). “Fitting a mixture model by expectation maximization to discover motifs in biopolymers.” Proceedings of the Second International Conference on Intelligent Systems for Molecular Biology, 2, 28--36.

See also

Examples

# NOT RUN {
## To check that you are properly linking to the binary:
run_meme()
# }