Shuffle motifs by column.

Given a set of motifs, shuffle the columns between them. Currently does not support keeping the 'multifreq' slot. Only the 'bkg', 'nsites', 'strand', and 'bkgsites' slots will be preserved. Uses the same shuffling methods as shuffle_sequences(). When shuffling more than one motif, they are shuffled together.

shuffle_motifs(motifs, k = 2, method = "linear", leftovers = "asis")

Arguments

motifs
See convert_motifs() for acceptable formats.

k
numeric(1) K-let size.

method
character(1) One of c('linear', 'random'). Only relevant if k > 1. See details.

leftovers
character(1) For method = 'random'. One of c('asis', 'first', 'split', 'discard'). See details.

Value

Motifs. See convert_motifs() for available output formats.

Details

If method = 'linear', then the input positions are split linearly every k columns after which they are shuffled. If method = random, then sets of k-columns are chosen randonmly before being shuffled. This leaves leftover column islands smaller than k; these can be left asis, placed first, split between the beginning and the end, or discarded. See shuffle_motifs().

See also