universalmotif: Motif class.

Container for motif objects. See create_motif() for creating motifs as well as a more detailed description of the slots. For a brief description of available methods, see examples.

# S4 method for universalmotif
[(x, i)

# S4 method for universalmotif,ANY,ANY,ANY
[(x, i) <- value

# S4 method for universalmotif
initialize(.Object, name, altname, family,
  organism, motif, alphabet = "DNA", type, icscore, nsites,
  pseudocount = 0.8, bkg, bkgsites, consensus, strand = "+-", pval,
  qval, eval, multifreq, extrainfo)

# S4 method for universalmotif
show(object)

# S4 method for universalmotif
as.data.frame(x)

# S4 method for universalmotif
subset(x, select)

# S4 method for universalmotif
normalize(object)

# S4 method for universalmotif
rowMeans(x)

# S4 method for universalmotif
colMeans(x)

# S4 method for universalmotif
colSums(x)

# S4 method for universalmotif
rowSums(x)

# S4 method for universalmotif
nrow(x)

# S4 method for universalmotif
ncol(x)

# S4 method for universalmotif
colnames(x)

# S4 method for universalmotif
rownames(x)

# S4 method for universalmotif
cbind(..., deparse.level = 0)

Arguments

x
universalmotif Motif.

i
character Slot.

value
Object to replace slot with.

.Object
universalmotif Final motif.

name
character(1) Motif name.

altname
character(1) Alternate motif name.

family
character(1) Transcription factor family.

organism
character(1) Species of origin.

motif
matrix Each column represents a position in the motif.

alphabet
character(1) One of c('DNA', 'RNA', 'AA', 'custom'), or a combined string representing the letters.

type
character(1) One of c('PCM', 'PPM', 'PWM', 'ICM').

icscore
numeric(1) Total information content. Automatically generated.

nsites
numeric(1) Number of sites the motif was constructed from.

pseudocount
numeric(1) Correction to be applied to prevent -Inf from apearing in PWM matrices.

bkg
numeric Must sum to 1 and be equal in length to the alphabet length. If missing, assumes a uniform background.

bkgsites
numeric(1) Total number of sites used to find the motif.

consensus
character(1) Consensus string. Automatically generated for 'DNA', 'RNA', and 'AA' alphabets.

strand
character(1) Whether the motif is specific to a certain strand.

pval
numeric(1) P-value associated with motif.

qval
numeric(1) Adjusted P-value associated with motif.

eval
numeric(1) E-value associated with motif.

multifreq
list See add_multifreq().

extrainfo
character Any other extra information, represented as a named character vector.

object
universalmotif Motif.

select
numeric Columns to keep.

...
universalmotif Motifs.

deparse.level
Unused.

Value

A motif object of class universalmotif.

Slots

name

character(1)

altname

character(1)

family

character(1)

organism

character(1)

motif

matrix

alphabet

character(1)

type

character(1)

icscore

numeric(1) Generated automatically.

nsites

numeric(1)

pseudocount

numeric(1)

bkg

numeric Length equal to number of letters in alphabet.

bkgsites

numeric(1)

consensus

character Generated automatically.

strand

character(1)

pval

numeric(1)

qval

numeric(1)

eval

numeric(1)

multifreq

list

extrainfo

character

Examples

## [
## Access the slots.
motif <- create_motif()
motif["motif"]

#>             T            K          A            A            G          A
#> A 0.036685844 1.437478e-02 0.68464418 7.472794e-01 2.333391e-01 0.68905484
#> C 0.001499805 5.277781e-05 0.01056073 1.620201e-07 3.880123e-05 0.08855173
#> G 0.019098916 4.190816e-01 0.17532289 2.424517e-01 7.666187e-01 0.03148839
#> T 0.942715435 5.664909e-01 0.12947220 1.026871e-02 3.465216e-06 0.19090504
#>            C          G           T            M
#> A 0.04855025 0.11608742 0.016068184 0.5062549789
#> C 0.68880582 0.06053184 0.061263489 0.4905346505
#> G 0.11394730 0.74430372 0.002726793 0.0027164485
#> T 0.14869664 0.07907702 0.919941534 0.0004939221

# you can also access multiple slots at once, released as a list
motif[c("motif", "name")]

#> $motif
#>             T            K          A            A            G          A
#> A 0.036685844 1.437478e-02 0.68464418 7.472794e-01 2.333391e-01 0.68905484
#> C 0.001499805 5.277781e-05 0.01056073 1.620201e-07 3.880123e-05 0.08855173
#> G 0.019098916 4.190816e-01 0.17532289 2.424517e-01 7.666187e-01 0.03148839
#> T 0.942715435 5.664909e-01 0.12947220 1.026871e-02 3.465216e-06 0.19090504
#>            C          G           T            M
#> A 0.04855025 0.11608742 0.016068184 0.5062549789
#> C 0.68880582 0.06053184 0.061263489 0.4905346505
#> G 0.11394730 0.74430372 0.002726793 0.0027164485
#> T 0.14869664 0.07907702 0.919941534 0.0004939221
#> 
#> $name
#> [1] "motif"
#> 

## [<-
## Replace the slots.
motif["name"] <- "new name"
# some slots are protected
# motif["consensus"] <- "AAAA"  # not allowed

## c
## Assemble a list of motifs.
c(motif, motif)

#> [[1]]
#> 
#>        Motif name:   new name
#>          Alphabet:   DNA
#>              Type:   PPM
#>           Strands:   +-
#>          Total IC:   10.25462
#>         Consensus:   TKAAGACGTM
#>      Target sites:   213
#> 
#>      T    K    A    A    G    A    C    G    T    M
#> A 0.04 0.01 0.68 0.75 0.23 0.69 0.05 0.12 0.02 0.51
#> C 0.00 0.00 0.01 0.00 0.00 0.09 0.69 0.06 0.06 0.49
#> G 0.02 0.42 0.18 0.24 0.77 0.03 0.11 0.74 0.00 0.00
#> T 0.94 0.57 0.13 0.01 0.00 0.19 0.15 0.08 0.92 0.00
#> 
#> [[2]]
#> 
#>        Motif name:   new name
#>          Alphabet:   DNA
#>              Type:   PPM
#>           Strands:   +-
#>          Total IC:   10.25462
#>         Consensus:   TKAAGACGTM
#>      Target sites:   213
#> 
#>      T    K    A    A    G    A    C    G    T    M
#> A 0.04 0.01 0.68 0.75 0.23 0.69 0.05 0.12 0.02 0.51
#> C 0.00 0.00 0.01 0.00 0.00 0.09 0.69 0.06 0.06 0.49
#> G 0.02 0.42 0.18 0.24 0.77 0.03 0.11 0.74 0.00 0.00
#> T 0.94 0.57 0.13 0.01 0.00 0.19 0.15 0.08 0.92 0.00
#> 

## as.data.frame
## Represent a motif as a data.frame. The actual motif matrix is lost.
## Necessary for `summarise_motifs`.
as.data.frame(motif)

#>       name altname family organism alphabet  icscore nsites bkgsites  consensus
#> 1 new name    <NA>   <NA>     <NA>      DNA 10.25462    213       NA TKAAGACGTM
#>   strand pval qval eval
#> 1     +-   NA   NA   NA

## subset
## Subset a motif matrix by column.
subset(motif, 3:7)  # extract motif core

#> 
#>        Motif name:   new name
#>          Alphabet:   DNA
#>              Type:   PPM
#>           Strands:   +-
#>          Total IC:   4.432664
#>         Consensus:   AAGAC
#>      Target sites:   213
#> 
#>      A    A    G    A    C
#> A 0.68 0.75 0.23 0.69 0.05
#> C 0.01 0.00 0.00 0.09 0.69
#> G 0.18 0.24 0.77 0.03 0.11
#> T 0.13 0.01 0.00 0.19 0.15

## normalize
## Apply the pseudocount slot (or `1`, if the slot is set to zero) to the
## motif matrix.
motif2 <- create_motif("AAAAA", nsites = 100, pseudocount = 1)
normalize(motif2)

#> 
#>        Motif name:   motif
#>          Alphabet:   DNA
#>              Type:   PPM
#>           Strands:   +-
#>          Total IC:   9.625166
#>         Consensus:   AAAAA
#>      Target sites:   100
#> 
#>      A    A    A    A    A
#> A 0.99 0.99 0.99 0.99 0.99
#> C 0.00 0.00 0.00 0.00 0.00
#> G 0.00 0.00 0.00 0.00 0.00
#> T 0.00 0.00 0.00 0.00 0.00

## rowMeans
## Calculate motif rowMeans.
rowMeans(motif)

#>         A         C         G         T 
#> 0.3092339 0.1401840 0.2517756 0.2988065 

## colMeans
## Calculate motif colMeans.
colMeans(motif)

#>    T    K    A    A    G    A    C    G    T    M 
#> 0.25 0.25 0.25 0.25 0.25 0.25 0.25 0.25 0.25 0.25 

## colSums
## Calculate motif colSums
colSums(motif)

#> T K A A G A C G T M 
#> 1 1 1 1 1 1 1 1 1 1 

## rowSums
## Calculate motif rowSums.
rowSums(motif)

#>        A        C        G        T 
#> 3.092339 1.401840 2.517756 2.988065 

## nrow
## Count motif rows.
nrow(motif)

#> [1] 4

## ncol
## Count motif columns.
ncol(motif)

#> [1] 10

## colnames
## Get motif colnames.
colnames(motif)

#>  [1] "T" "K" "A" "A" "G" "A" "C" "G" "T" "M"

## rownames
## Get motif rownames.
rownames(motif)

#> [1] "A" "C" "G" "T"

## cbind
## Bind motifs together to create a new motif.
cbind(motif, motif2)

#> 
#>        Motif name:   new name/motif
#>          Alphabet:   DNA
#>              Type:   PPM
#>           Strands:   +-
#>          Total IC:   19.37954
#>         Consensus:   TKAAGACGTMAAAAA
#>      Target sites:   213
#> 
#>      T    K    A    A    G    A    C    G    T    M    A    A    A    A    A
#> A 0.04 0.01 0.68 0.75 0.23 0.69 0.05 0.12 0.02 0.51 0.99 0.99 0.99 0.99 0.99
#> C 0.00 0.00 0.01 0.00 0.00 0.09 0.69 0.06 0.06 0.49 0.00 0.00 0.00 0.00 0.00
#> G 0.02 0.42 0.18 0.24 0.77 0.03 0.11 0.74 0.00 0.00 0.00 0.00 0.00 0.00 0.00
#> T 0.94 0.57 0.13 0.01 0.00 0.19 0.15 0.08 0.92 0.00 0.00 0.00 0.00 0.00 0.00